+++ .esbx/plugin/static.ts
@@ -1,6 +1,7 @@
 import type {Plugin} from 'esbuild'
-import fs from 'fs-extra'
+import fs from 'fs'
 import path from 'path'
+import {copyDir} from '../copy'
 
 export const staticPlugin: Plugin = {
   name: '@esbx/static',
@@ -31,7 +32,7 @@ export const staticPlugin: Plugin = {
         const output = source.replace('packages' + path.sep, 'dist' + path.sep)
         if (fs.existsSync(source)) {
           if (!fs.existsSync(output)) fs.mkdirSync(output, {recursive: true})
-          const task = fs.copy(source, output, {overwrite: true})
+          const task = copyDir(source, output)
           tasks.push(task)
         }
       }