+++ apps/web/src/pages/index.tsx
@@ -4,7 +4,9 @@ import {PageView} from '../view/PageView'
 import {pageViewQuery} from '../view/PageView.server'
 
 export async function getStaticProps(context: GetStaticPropsContext) {
-  const pages = backend.loadPages('web', context.previewData as string)
+  const pages = backend.loadPages('web', {
+    previewToken: context.previewData as string
+  })
   const paths = (context.params?.slug as Array<string>) || []
   const slug = '/' + paths.join('/')
   const page = await pageViewQuery(pages, slug)