+++ packages/backend/src/router/Router.ts
@@ -1,21 +1,6 @@
 import {Outcome} from '@alinea/core/Outcome'
-import {
-  Headers as WebHeaders,
-  Request as WebRequest,
-  Response as WebResponse
-} from '@remix-run/web-fetch'
-import {File} from '@remix-run/web-file'
+import {Request, Response} from '@alinea/iso'
 import {parse} from 'regexparam'
-
-export type Request = globalThis.Request
-export const Request = WebRequest
-export type Response = globalThis.Response
-export const Response = WebResponse
-export type Headers = globalThis.Headers
-export const Headers = WebHeaders
-
-// web-std/io#63
-globalThis.File = File
 
 export type Handle<In, Out> = {
   (input: In): Out | undefined | Promise<Out | undefined>