+++ packages/cli/src/Serve.ts
@@ -39,6 +39,7 @@ export async function serve(options: ServeOptions) {
   const storeLocation = path.join(outDir, 'store.js')
   const genConfigFile = path.join(outDir, 'config.js')
   const backendFile = path.join(outDir, 'backend.js')
+  const draftsFile = path.join(outDir, 'drafts.js')
   const clients: Array<ServerResponse> = []
   const {version} = require('react/package.json')
   const isReact18 = semver.compare(version, '18.0.0', '>=')
@@ -92,7 +93,6 @@ export async function serve(options: ServeOptions) {
   app.get('/', (req, res) => {
     res.send(`<!DOCTYPE html>
     <meta charset="utf-8" />
-    <title>Alinea</title>
     <link rel="icon" href="data:," />
     <link href="./config.css" rel="stylesheet" />
     <link href="./entry.css" rel="stylesheet" />
@@ -193,6 +193,7 @@ export async function serve(options: ServeOptions) {
 
   async function devServer() {
     const unique = Date.now()
+    const {createStore: createDraftStore} = await import(`file://${draftsFile}`)
     let backend
     if (production) {
       backend = (await import(`file://${backendFile}`)).backend
@@ -205,7 +206,8 @@ export async function serve(options: ServeOptions) {
         config,
         createStore,
         port,
-        cwd
+        cwd,
+        createDraftStore
       })
     }
     return nodeHandler(backend.handle)