+++ packages/cli/src/static/workspace/pages.js
@@ -1,5 +1,0 @@
-import {Pages} from '@alinea/backend'
-import {createStore} from '../store.js'
-import {workspace} from './schema.js'
-
-export const pages = new Pages(workspace, createStore)