+++ packages/cli/src/util/EnsureEnv.ts
@@ -1,0 +1,7 @@
+import dotenv from 'dotenv'
+import findConfig from 'find-config'
+
+export function ensureEnv(cwd: string) {
+  const path = findConfig('.env', {cwd})
+  if (path) dotenv.config({path})
+}