+++ packages/env.d.ts
@@ -1,13 +1,9 @@
-declare module '.alinea'
-
 declare module '*.module.scss' {
   const classes: {[key: string]: string}
   export default classes
 }
 
 declare module 'eval:*'
-declare module 'get-random-values'
-declare module 'leb128'
 declare module 'postcss-pxtorem'
 // @types/react-helmet depends on react < 18 types which breaks things
 declare module 'react-helmet'