+++ packages/input/richtext/src/RichTextToolbar.tsx
@@ -19,6 +19,7 @@ import {IcRoundFormatItalic} from '@alinea/ui/icons/IcRoundFormatItalic'
 import {IcRoundFormatListBulleted} from '@alinea/ui/icons/IcRoundFormatListBulleted'
 import {IcRoundFormatListNumbered} from '@alinea/ui/icons/IcRoundFormatListNumbered'
 import {IcRoundLink} from '@alinea/ui/icons/IcRoundLink'
+import {IcRoundQuote} from '@alinea/ui/icons/IcRoundQuote'
 import {IcRoundUnfoldMore} from '@alinea/ui/icons/IcRoundUnfoldMore'
 /*import {IcRoundRedo} from '@alinea/ui/icons/IcRoundRedo'
 import {IcRoundUndo} from '@alinea/ui/icons/IcRoundUndo'*/
@@ -276,6 +277,17 @@ export const RichTextToolbar = forwardRef(function RichTextToolbar(
             title="Link"
             onClick={handleLink}
             active={editor.isActive('link')}
+          />
+          <div className={styles.root.separator()} />
+          <IconButton
+            icon={IcRoundQuote}
+            size={18}
+            title="Blockquote"
+            onClick={e => {
+              e.preventDefault()
+              exec().toggleBlockquote().run()
+            }}
+            active={editor.isActive('blockquote')}
           />
         </HStack>
       </div>