+++ changelog.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+## [0.1.4]
+
+- Replace CJS dependencies to pure ESM (#161)
+- Added blockquote option to the rich text field
+
+## [0.1.3]
+
+- The drafts of the default development backend are now placed with the
+  generated code which means draft changes will result in browser reloads.
+
 ## [0.1.2]
 
 - A RichText react component is exposed from @alinea/ui to render rich text